requirements.txt
fastapi
uvicorn
python-dotenv
langchain_google_genai
pydantic
sqlalchemy
jinja2
python-multipart
requests
opencv-python
